requirements.txt
amqp==2.5.0
anyjson==0.3.3
asgiref==3.1.2
billiard==3.5.0.5
celery==4.1.1
channels==2.2.0
channels-redis==2.4.0
Django==2.2.1
django-celery==3.3.0
django-celery-beat==1.5.0
django-cors-headers==3.0.2
django-timezone-field==3.0
djangorestframework==3.9.4
kombu==4.6.3
python-crontab==2.3.6
python-dateutil==2.8.0
python-magic==0.4.15
python-magic-bin==0.4.14
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
